[feat]#187 AdditionViewModel 두 개로 구분하여 작업
